+++ plugin/src/test/kotlin/org/jlleitschuh/gradle/ktlint/AbstractPluginTest.kt
@@ -66,6 +66,12 @@ abstract class AbstractPluginTest {
     protected
     fun File.withFailingSources() = createSourceFile("src/main/kotlin/fail-source.kt", """val  foo    =     "bar"""")
 
+    protected fun File.restoreFailingSources() {
+        val sourceFile = resolve("src/main/kotlin/fail-source.kt")
+        sourceFile.delete()
+        withFailingSources()
+    }
+
     protected
     fun File.withAlternativeFailingSources(baseDir: String) =
         createSourceFile("$baseDir/fail-source.kt", """val  foo    =     "bar"""")